Split formatter into separate processors
